requirements.txt
# Core Framework
streamlit>=1.28.0

# AI/ML Libraries
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
langchain>=0.1.0
ollama>=0.1.0
google-generativeai>=0.3.0

# Document Processing
pypdf>=3.17.0

# Data Processing & Visualization
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0

# Web Framework (for WhatsApp bot)
flask>=3.0.0
twilio>=8.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Translation (for 22 Indian languages)
# Using deep-translator instead of googletrans (no conflicts, actively maintained)
deep-translator>=1.11.0
